+++ src/main/java/io/vacco/jtinn/JtUpdate.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class JtUpdate {
 
   public interface JtUpdater extends Serializable {
-    void apply(double[] lm1a, JtLayers.JtLayer l);
+    void apply(float[] lm1a, JtLayers.JtLayer l);
     void onEpochEnd(int epoch);
   }
 
@@ -13,18 +13,18 @@ public class JtUpdate {
 
     private static final long serialVersionUID = JtUtil.version;
 
-    public double learningRate, annealingRate;
+    public float learningRate, annealingRate;
 
-    public JtSgdUpdater init(double learningRate, double annealingRate) {
+    public JtSgdUpdater init(float learningRate, float annealingRate) {
       this.learningRate = learningRate;
       this.annealingRate = annealingRate;
       return this;
     }
 
-    @Override public void apply(double[] lm1a, JtLayers.JtLayer l) {
+    @Override public void apply(float[] lm1a, JtLayers.JtLayer l) {
       for (int j = 0; j < l.size(); j++) {
         for (int w = 0; w < l.weightSize(); w++) {
-          double pda = l instanceof JtLayers.JtOutputLayer ? lm1a[w] : l.actFn.pd(l.a[j]) * lm1a[w];
+          float pda = l instanceof JtLayers.JtOutputLayer ? lm1a[w] : l.actFn.pd(l.a[j]) * lm1a[w];
           l.w[j][w] = l.w[j][w] - (learningRate * l.δ[j] * pda);
         }
         l.b[j] = l.b[j] - l.δ[j];
